My current CLAUDE.local.md - WIP
The term Prompt Engineering needs to be banned. But that is what one would
probably describe this as.

This one is too specific to my personal development setup and does prove
helpful as I no longer have to repeat a bunch of basic instructions.

A more distilled generic one could make sense as a top level CLAUDE.md file to
make it work better for everyone using Claude Code in the cpython repo.

Expect these to evolve as models do.  Claude 4 Sonnet and Opus deal with our
codebase a lot better than Sonnet 3.6 and 3.7, yet we're still a challenge.